interrupts: provide system PIC registration and dispatch
The GIC functionality we've had to date has been a very leaky
abstraction, spilling knowledge of the GIC (and of the FDT) throughout
`unix`. With the work that's been done to isolate firmware knowledge
to the GIC drivers, combined with the work to attach GIC drivers using
standard mechanisms, we have the opportunity to clean up our
abstractions a little and deal with the amount of global state we've
accrued.

Introduce `syspic`, which is a hub layer for dealing with the system
programmable interrupt controller. This layer subsumes the interrupt
state tracking functionality implemented for debugger support, provides
a registration interface for the root interrupt controller and provides
dispatch functions to per-driver callback functions, passing any context
registered by the driver on otherwise context-free calls to the driver.

The interfaces consumed by `unix` are documented in `syspic.h`, while
the driver and debugger integration interfaces are documented in
`syspic_impl.h`.

Requirements for the state tracking locking protocol are documented
in `syspic_impl.h` and includes guidance on integrating DDI-based
interrupts as well as SGIs.

While here, do a quick drive-by fix on `arm_gtmr` so that the IPI
registered there is correctly attributed to the driver.