address-cells should default to 0 for interrupt-map
The standard states:
Both the child node and the interrupt parent node are required
to have #address-cells and #interrupt-cells properties defined.
If a unit address component is not required, #address-cells
shall be explicitly defined to be zero.

However, we've seen interrup mappins on the RPI4, where #address-cells
is not defined in the parent (a calculation reveals it should be 0).
If we propagate to the parent's parent we'd likely end up with a wrong number.