pci_common: make `pci_config_rd*`/`pci_config_wr*` private
These are used to drive the DDI access to PCI config space, via being set into
the access handle via `pci_fm_acc_setup`, and use `pci_get?_func` in their
implementation.

They're not meant to be used outside of that environment, and aren't used
outside of that environment, so let's pin that down.